api/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5
billiard==3.6.4.0
black==20.8b1
celery==5.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.22
django-extensions==3.1.2
django-imagekit==4.0.2
django-nine==0.2.4
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
drf-extensions==0.7.0
drf-yasg==1.20.0
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
filelock==3.0.12
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
kombu==5.1.0
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.2
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.3
pathspec==0.8.1
pilkit==2.0
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
ptvsd==4.3.2
py==1.10.0
pycodestyle==2.6.0
PyJWT==2.0.1
pylint==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-logstash==0.4.6
pytz==2021.1
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.23.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
virtualenv==20.4.4
wcwidth==0.2.5
wrapt==1.12.1
